requirements.txt
requests
nfcpy
gql
